Remove batch-size Settings
