+++ core/__tests__/tasks/profile/enqueueExports/enqueueExports.ts
@@ -1,21 +1,11 @@
 import { helper } from "@grouparoo/spec-helper";
 import { api, task, specHelper } from "actionhero";
-import {
-  Import,
-  plugin,
-  Profile,
-  ProfileProperty,
-  Property,
-} from "../../../src";
+import { Import, Profile, ProfileProperty, Property } from "../../../../src";
 
 describe("tasks/profiles:enqueueExports", () => {
   helper.grouparooTestServer({ truncate: true, enableTestPlugin: true });
   beforeEach(async () => await api.resque.queue.connection.redis.flushdb());
   beforeAll(async () => await helper.factories.properties());
-
-  afterEach(async () => {
-    await plugin.updateSetting("core", "runs-profile-batch-size", 100);
-  });
 
   describe("profiles:enqueueExports", () => {
     test("can be enqueued", async () => {
@@ -140,46 +130,6 @@ describe("tasks/profiles:enqueueExports", () => {
       await mario.destroy();
       await luigi.destroy();
       await toad.destroy();
-    });
-
-    test("batch size can be configured with a setting", async () => {
-      await plugin.updateSetting("core", "runs-profile-batch-size", 1);
-
-      const mario: Profile = await helper.factories.profile();
-      await mario.import();
-      await mario.update({ state: "ready" });
-      const marioImport: Import = await helper.factories.import(
-        null,
-        {},
-        mario.id
-      );
-      await marioImport.update({
-        groupsUpdatedAt: new Date(),
-        profileUpdatedAt: new Date(),
-        exportedAt: null,
-      });
-
-      const luigi: Profile = await helper.factories.profile();
-      await luigi.import();
-      await luigi.update({ state: "ready" });
-      const luigiImport: Import = await helper.factories.import(
-        null,
-        {},
-        luigi.id
-      );
-      await luigiImport.update({
-        groupsUpdatedAt: new Date(),
-        profileUpdatedAt: new Date(),
-        exportedAt: null,
-      });
-
-      await specHelper.runTask("profiles:enqueueExports", {});
-
-      const foundTasks = await specHelper.findEnqueuedTasks("profile:export");
-      expect(foundTasks.length).toEqual(1);
-
-      await mario.destroy();
-      await luigi.destroy();
     });
   });
 });
