+++ core/src/tasks/group/exportToCSV.ts
@@ -1,6 +1,5 @@
-import { api } from "actionhero";
+import { api, config } from "actionhero";
 import { Group } from "../../models/Group";
-import { plugin } from "../../modules/plugin";
 import { groupExportToCSV } from "../../modules/groupExport";
 import { CLSTask } from "../../classes/tasks/clsTask";
 
@@ -19,11 +18,7 @@ export class GroupExportToCSV extends CLSTask {
   }
 
   async runWithinTransaction(params) {
-    const limit: number =
-      params.limit ||
-      parseInt(
-        (await plugin.readSetting("core", "runs-profile-batch-size")).value
-      );
+    const limit: number = params.limit || config.batchSize.imports;
 
     const group = await Group.findById(params.groupId);
