+++ plugins/@grouparoo/mailchimp/__tests__/integration/mailchimp-import.ts
@@ -59,10 +59,6 @@ describe("integration/runs/mailchimp-import", () => {
   });
 
   beforeAll(async () => {
-    await plugin.updateSetting("core", "runs-profile-batch-size", 100);
-  });
-
-  beforeAll(async () => {
     await helper.factories.properties();
     helper.disableTestPluginImport();
   });
